src/requirements.txt
auto_mix_prep==0.2.0
fastapi==0.110.0
keras==2.15.0
matplotlib==3.8.2
mlflow==2.10.1
nltk==3.8.1
numpy==1.26.4
pandas==2.2.1
pydantic==2.6.2
PyYAML==6.0.1
scikit_learn==1.4.0
tensorflow-cpu==2.15.0
transformers==4.37.2
uvicorn==0.27.1
